# source/ProductOrderCaptureAndValidation/productOrderingMicroservice/implementation/api/swagger.yaml
swagger: '2.0'
info:
  title: Product Ordering
  description: "**TMF API Reference : TMF 622 - Product Ordering Management**\n\n\
    **Release : 19.0 - June 2019**\n\nThe Product Ordering API provides a standardized\
    \ mechanism for placing a product order with all of the necessary order parameters.\
    \ The API consists of a simple set of operations that interact with CRM/Order\
    \ Negotiation systems in a consistent manner. A product order is created based\
    \ on a product offer that is defined in a catalog. The product offer identifies\
    \ the product or set of products that are available to a customer, and includes\
    \ characteristics such as pricing, product options and market. This API provide\
    \ a task based resource to request order cancellation.\n\nThe product order references\
    \ the product offer and identifies any specific requests made by the customer.\n\
    \n**Product Order resource**\nA Product Order is a type of order which can be\
    \ used to place an order between a customer and a service provider or between\
    \ a service provider and a partner and vice versa. Main Product Order attributes\
    \ are its identifier, state, priority category (mass market, Enterprise, etc.)\
    \ related dates (start, completion, etc.), related billing account, related parties\
    \ and order items. Main Order Items (aka order lines) attributes are the ordered\
...
    \n\nCopyright � TM Forum 2019. All Rights Reserved\n\n\n"
  version: 4.0.0
basePath: /tmf-api/productOrderingManagement/v4/
schemes:
...
consumes:
...
produces:
...
paths:
...
  /listener/productOrderCreateEvent:
    post:
      operationId: listenToProductOrderCreateEvent
      summary: Client listener for entity ProductOrderCreateEvent
      description: Example of a client listener for receiving the notification ProductOrderCreateEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/ProductOrderCreateEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/productOrderAttributeValueChangeEvent:
    post:
      operationId: listenToProductOrderAttributeValueChangeEvent
      summary: Client listener for entity ProductOrderAttributeValueChangeEvent
      description: Example of a client listener for receiving the notification ProductOrderAttributeValueChangeEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/ProductOrderAttributeValueChangeEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/productOrderDeleteEvent:
    post:
      operationId: listenToProductOrderDeleteEvent
      summary: Client listener for entity ProductOrderDeleteEvent
      description: Example of a client listener for receiving the notification ProductOrderDeleteEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/ProductOrderDeleteEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/productOrderStateChangeEvent:
    post:
      operationId: listenToProductOrderStateChangeEvent
      summary: Client listener for entity ProductOrderStateChangeEvent
      description: Example of a client listener for receiving the notification ProductOrderStateChangeEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/ProductOrderStateChangeEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/productOrderInformationRequiredEvent:
    post:
      operationId: listenToProductOrderInformationRequiredEvent
      summary: Client listener for entity ProductOrderInformationRequiredEvent
      description: Example of a client listener for receiving the notification ProductOrderInformationRequiredEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/ProductOrderInformationRequiredEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/cancelProductOrderCreateEvent:
    post:
      operationId: listenToCancelProductOrderCreateEvent
      summary: Client listener for entity CancelProductOrderCreateEvent
      description: Example of a client listener for receiving the notification CancelProductOrderCreateEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/CancelProductOrderCreateEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/cancelProductOrderStateChangeEvent:
    post:
      operationId: listenToCancelProductOrderStateChangeEvent
      summary: Client listener for entity CancelProductOrderStateChangeEvent
      description: Example of a client listener for receiving the notification CancelProductOrderStateChangeEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/CancelProductOrderStateChangeEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
  /listener/cancelProductOrderInformationRequiredEvent:
    post:
      operationId: listenToCancelProductOrderInformationRequiredEvent
      summary: Client listener for entity CancelProductOrderInformationRequiredEvent
      description: Example of a client listener for receiving the notification CancelProductOrderInformationRequiredEvent
      tags:
      - notification listeners (client side)
      parameters:
      - name: data
        required: true
        in: body
        description: The event data
        schema:
          $ref: '#/definitions/CancelProductOrderInformationRequiredEvent'
      responses:
        '201':
          description: Notified
          schema:
            $ref: '#/definitions/EventSubscription'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: Method Not allowed
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      x-swagger-router-controller: NotificationListenersClientSide
definitions:
...
  AgreementRef:
    type: object
    description: Agreement reference. An agreement represents a contract or arrangement,
      either written or verbal and sometimes enforceable by law, such as a service
      level agreement or a customer price agreement. An agreement involves a number
      of other business entities, such as products, services, and resources and/or
      their specifications.
    required:
    - id
    properties:
      id:
        type: string
        description: Unique identifier of a related entity.
      href:
        type: string
        description: Reference of the related entity.
      name:
        type: string
        description: Name of the agreement
      '@baseType':
        type: string
        description: When sub-classing, this defines the super-class
      '@schemaLocation':
        type: string
        format: uri
        description: A URI to a JSON-Schema file that defines additional attributes
          and relationships
      '@type':
        type: string
        description: When sub-classing, this defines the sub-class entity name
      '@referredType':
        type: string
        description: The actual type of the target instance when needed for disambiguation.
...
  ProductOrder:
    type: object
    description: A Product Order is a type of order which  can  be used to place an
      order between a customer and a service provider or between a service provider
      and a partner and vice versa,
    required:
    - productOrderItem
    properties:
      id:
        type: string
        description: ID created on repository side (OM system)
      href:
        type: string
        description: Hyperlink to access the order
      cancellationDate:
        type: string
        format: date-time
        description: 'Date when the order is cancelled. This is used when order is
          cancelled. '
      cancellationReason:
        type: string
        description: 'Reason why the order is cancelled. This is used when order is
          cancelled. '
      category:
        type: string
        description: Used to categorize the order from a business perspective that
          can be useful for the OM system (e.g. "enterprise", "residential", ...)
      completionDate:
        type: string
        format: date-time
        description: Date when the order was completed
      description:
        type: string
        description: Description of the product order
      expectedCompletionDate:
        type: string
        format: date-time
        description: Expected delivery date amended by the provider
      externalId:
        type: string
        description: ID given by the consumer and only understandable by him (to facilitate
          his searches afterwards)
      notificationContact:
        type: string
        description: Contact attached to the order to send back information regarding
          this order
      orderDate:
        type: string
        format: date-time
        description: Date when the order was created
      priority:
        type: string
        description: 'A way that can be used by consumers to prioritize orders in
          OM system (from 0 to 4 : 0 is the highest priority, and 4 the lowest)'
      requestedCompletionDate:
        type: string
        format: date-time
        description: Requested delivery date from the requestor perspective
      requestedStartDate:
        type: string
        format: date-time
        description: 'Order fulfillment start date wished by the requestor. This is
          used when, for any reason, requestor cannot allow seller to begin to operationally
          begin the fulfillment before a date. '
      agreement:
        type: array
        items:
          $ref: '#/definitions/AgreementRef'
        description: A reference to an agreement defined in the context of the product
          order
      billingAccount:
        $ref: '#/definitions/BillingAccountRef'
      channel:
        type: array
        items:
          $ref: '#/definitions/RelatedChannel'
      note:
        type: array
        items:
          $ref: '#/definitions/Note'
      orderTotalPrice:
        type: array
        items:
          $ref: '#/definitions/OrderPrice'
      payment:
        type: array
        items:
          $ref: '#/definitions/PaymentRef'
      productOfferingQualification:
        type: array
        items:
          $ref: '#/definitions/ProductOfferingQualificationRef'
      productOrderItem:
        type: array
        items:
          $ref: '#/definitions/ProductOrderItem'
        minItems: 1
      quote:
        type: array
        items:
          $ref: '#/definitions/QuoteRef'
      relatedParty:
        type: array
        items:
          $ref: '#/definitions/RelatedParty'
      state:
        $ref: '#/definitions/ProductOrderStateType'
        description: Tracks the lifecycle status of the product order, such as Acknowledged,
          Rejected, InProgress, Pending and so on.
      '@baseType':
        type: string
        description: When sub-classing, this defines the super-class
      '@schemaLocation':
        type: string
        format: uri
        description: A URI to a JSON-Schema file that defines additional attributes
          and relationships
      '@type':
        type: string
        description: When sub-classing, this defines the sub-class entity name
...
  EventSubscription:
    type: object
    description: Sets the communication endpoint address the service instance must
      use to deliver notification information
    required:
    - id
    - callback
    properties:
      id:
        type: string
        description: Id of the listener
      callback:
        type: string
        description: The callback being registered.
      query:
        type: string
        description: additional data to be passed
...
  ProductOrderCreateEvent:
    type: object
    description: The notification data structure
    properties:
      id:
        type: string
        description: Identifier of the Process flow
      href:
        type: string
        description: Reference of the ProcessFlow
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/ProductOrderCreateEventPayload'
  ProductOrderCreateEventPayload:
    type: object
    description: The event data structure
    properties:
      productOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/ProductOrder'
  ProductOrderAttributeValueChangeEvent:
    type: object
    description: The notification data structure
    properties:
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      fieldPath:
        type: string
        description: The path identifying the object field concerned by this notification.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/ProductOrderAttributeValueChangeEventPayload'
  ProductOrderAttributeValueChangeEventPayload:
    type: object
    description: The event data structure
    properties:
      productOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/ProductOrder'
  ProductOrderDeleteEvent:
    type: object
    description: The notification data structure
    properties:
      id:
        type: string
        description: Identifier of the Process flow
      href:
        type: string
        description: Reference of the ProcessFlow
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/ProductOrderDeleteEventPayload'
  ProductOrderDeleteEventPayload:
    type: object
    description: The event data structure
    properties:
      productOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/ProductOrder'
  ProductOrderStateChangeEvent:
    type: object
    description: The notification data structure
    properties:
      id:
        type: string
        description: Identifier of the Process flow
      href:
        type: string
        description: Reference of the ProcessFlow
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/ProductOrderStateChangeEventPayload'
  ProductOrderStateChangeEventPayload:
    type: object
    description: The event data structure
    properties:
      productOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/ProductOrder'
  ProductOrderInformationRequiredEvent:
    type: object
    description: The notification data structure
    properties:
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      fieldPath:
        type: string
        description: The path identifying the object field concerned by this notification.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/ProductOrderInformationRequiredEventPayload'
  ProductOrderInformationRequiredEventPayload:
    type: object
    description: The event data structure
    properties:
      productOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/ProductOrder'
  CancelProductOrderCreateEvent:
    type: object
    description: The notification data structure
    properties:
      id:
        type: string
        description: Identifier of the Process flow
      href:
        type: string
        description: Reference of the ProcessFlow
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/CancelProductOrderCreateEventPayload'
  CancelProductOrderCreateEventPayload:
    type: object
    description: The event data structure
    properties:
      cancelProductOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/CancelProductOrder'
  CancelProductOrderStateChangeEvent:
    type: object
    description: The notification data structure
    properties:
      id:
        type: string
        description: Identifier of the Process flow
      href:
        type: string
        description: Reference of the ProcessFlow
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/CancelProductOrderStateChangeEventPayload'
  CancelProductOrderStateChangeEventPayload:
    type: object
    description: The event data structure
    properties:
      cancelProductOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/CancelProductOrder'
  CancelProductOrderInformationRequiredEvent:
    type: object
    description: The notification data structure
    properties:
      eventId:
        type: string
        description: The identifier of the notification.
      eventTime:
        type: string
        format: date-time
        description: Time of the event occurrence.
      eventType:
        type: string
        description: The type of the notification.
      correlationId:
        type: string
        description: The correlation id for this event.
      domain:
        type: string
        description: The domain of the event.
      title:
        type: string
        description: The title of the event.
      description:
        type: string
        description: An explnatory of the event.
      priority:
        type: string
        description: A priority.
      timeOcurred:
        type: string
        format: date-time
        description: The time the event occured.
      fieldPath:
        type: string
        description: The path identifying the object field concerned by this notification.
      event:
        description: The event payload linked to the involved resource object
        $ref: '#/definitions/CancelProductOrderInformationRequiredEventPayload'
  CancelProductOrderInformationRequiredEventPayload:
    type: object
    description: The event data structure
    properties:
      cancelProductOrder:
        description: The involved resource data for the event
        $ref: '#/definitions/CancelProductOrder'
  Error:
    description: Used when an API throws an Error, typically with a HTTP error response-code
      (3xx, 4xx, 5xx)
    type: object
    required:
    - code
    - reason
    properties:
      code:
        type: string
        description: Application relevant detail, defined in the API or a common list.
      reason:
        type: string
        description: Explanation of the reason for the error which can be shown to
          a client user.
      message:
        type: string
        description: More details and corrective actions related to the error which
          can be shown to a client user.
      status:
        type: string
        description: HTTP Error code extension
      referenceError:
        type: string
        format: uri
        description: URI of documentation describing the error.
      '@baseType':
        type: string
        description: When sub-classing, this defines the super-class.
      '@schemaLocation':
        type: string
        format: uri
        description: A URI to a JSON-Schema file that defines additional attributes
          and relationships
      '@type':
        type: string
        description: When sub-classing, this defines the sub-class entity name.